Added LineParser to give quick interface to read files Moved the loading of the Tiles + Tiles'settings from a tileset file. Added scaler offset for Cursor

+++ trunk/Sources/Engine/VTime.cpp
@@ -69,7 +69,7 @@ bool VTime :: canUpdate(void)
 {
 	int timeEllapsed = SDL_GetTicks() - lastUpdateTime;
 
-	if ( msSecondsBetweenTwoUpdate < timeEllapsed )
+	if ( static_cast<int>(msSecondsBetweenTwoUpdate) < timeEllapsed )
 	{
 		lastUpdateTime = SDL_GetTicks();
 		return true;
